Move promtool config/rules file checks to their own package so they can be used by other projects.
